+++ crates/chainio/clients/avsregistry/src/writer.rs
@@ -305,10 +305,8 @@ impl AvsRegistryChainWriter {
             })
             .collect();
 
-        let operator_id = FixedBytes::from(
-            operator_id_from_g1_pub_key(bls_key_pair.public_key())
-                .map_err(|_| AvsRegistryError::GetOperatorId)?,
-        );
+        let operator_id = operator_id_from_g1_pub_key(bls_key_pair.public_key())
+            .map_err(|_| AvsRegistryError::GetOperatorId)?;
 
         let churn_wallet = PrivateKeySigner::from_str(&churn_signer_private_key)
             .map_err(|_| AvsRegistryError::InvalidPrivateKey)?;