+++ crates/utils/src/lib.rs
@@ -1,4 +1,5 @@
 //! This module exports generated bindings.
+pub mod common;
 pub mod rewardsv2;
 pub mod slashing;